sh: add test for cmake, make and make doc
We need to add test if cmake, make and make doc failed.

+++ build_dev_version.sh
@@ -1,4 +1,5 @@
 #!/bin/sh
+
 COVERAGE_FOLDER_NAME=coverage
 coverage=1
 
@@ -12,17 +13,20 @@ if [ ! -d "build" ]; then
  mkdir build
 fi
 cd build
-cmake ../ -DCMAKE_BUILD_TYPE="Debug" -DWITH_COVERAGE=1 -DWITH_DOXYGEN=1
-make
-make doc
 
-rm -rf ../test_coverage
-lcov --base-directory . --directory . --zerocounters -q
-make test
-lcov --base-directory . --directory . -c -o project_test.info
-lcov --remove project_test.info "/usr*" -o project_test.info
-genhtml -o test_coverage -t "MAS test coverage" --num-spaces 4 project_test.info
+if cmake ../ -DCMAKE_BUILD_TYPE="Debug" -DWITH_COVERAGE=1 -DWITH_DOXYGEN=1 ; then
+  if make ; then
+    if make doc ; then
+      rm -rf ../test_coverage
+      lcov --base-directory . --directory . --zerocounters -q
+      make test
+      lcov --base-directory . --directory . -c -o project_test.info
+      lcov --remove project_test.info "/usr*" -o project_test.info
+      genhtml -o test_coverage -t "MAS test coverage" --num-spaces 4 project_test.info
+      mv ../build/test_coverage ../$COVERAGE_FOLDER_NAME
+      rm -rf ../build ../doxygen/html ../doxygen.conf
+    fi
+  fi
+fi
 
 cd ..
-mv build/test_coverage ./$COVERAGE_FOLDER_NAME
-rm -rf build doxygen/html doxygen.conf
